backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
Pillow>=10.0.0
packbits==0.6
pyusb>=1.2.1
qrcode[pil]>=7.4
python-barcode>=0.15.1
openpyxl>=3.1.0
python-multipart>=0.0.9
